+++ tests/test_function/test_parse_string.py
@@ -138,6 +138,19 @@ def test_parse_split() -> None:
     assert_that(result).is_equal_to(expected_result)
 
 
+def test_parse_starts_with() -> None:
+    # given
+    sub_dir_path = EXAMPLES_DIR / "starts_with"
+    file_path = sub_dir_path / "main.bicep"
+    expected_result = json.loads((sub_dir_path / "result.json").read_text())
+
+    # when
+    result = BicepParser(file_path=file_path).json()
+
+    # then
+    assert_that(result).is_equal_to(expected_result)
+
+
 def test_parse_string() -> None:
     # given
     sub_dir_path = EXAMPLES_DIR / "string"