+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support `skip` function
 - Support `managementGroup` function
 - Support `tenant` function
+- Support `startsWith` function
 
 ## 0.2.0 - 2022-02-13